109550158_HW5/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu117
torch 
torchvision 
torchaudio 
numpy
opencv-python
tqdm
Pillow
matplotlib
pandas
PyYAML